continued refactor of fixed width types for opcodes, index and oparg

+++ src/pyjion/types.h
@@ -32,6 +32,6 @@ using namespace std;
 
 typedef int16_t py_opcode;
 typedef int16_t py_oparg;
-typedef size_t py_opindex;
+typedef uint32_t py_opindex;
 
 #endif //PYJION_TYPES_H
